fix: fix an error of zip function
* What is the problem?
If `data`(this is a porp of LineChart) had one or more elements,
zip function threw an error("The length of a must be the same length of b").
This problem was caused by calling zip(acc, cur, [[]]).

* How to resolve
This problem was resolved by calling reduce method
in the condition that `data` have more or equal to two elements
by early returning corner cases.